# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: School Bus Tracking API
  version: 0.7.0
  description: |
    API for the School Bus Tracking admin web and the Android Driver and Parent apps.

    ## Conventions
    - Base path for business endpoints: `/api/v1`. JSON fields are `snake_case`.
    - Timestamps are RFC 3339 UTC (`2026-10-07T02:15:00Z`); calendar dates are `YYYY-MM-DD`.
    - Mobile numbers are accepted as `9876543210`, `+91 98765 43210`, etc. and always returned as `+919876543210`.
    - Success responses wrap the payload: `{"data": ...}`. Lists add
      `"meta": {"page", "page_size", "total"}` and accept `?page=1&page_size=20` (max 100).
    - Errors: `{"error": {"code", "message", "fields"}}`. Branch on `code`, show `message`.
    - Every response carries an `X-Request-ID` header; include it when reporting a bug.

    ## Authentication
    - **Driver and Parent apps:** `POST /auth/otp/send` then `POST /auth/otp/verify` with `app` = `driver` or `parent`.
      On the dev environment every code is `123456` (no SMS is sent).
    - **Admin web:** `POST /auth/login` (email + password). If the response has `mfa_required: true`,
      call `POST /auth/login/2fa` with the `mfa_token` and the authenticator code.
    - Send `Authorization: Bearer <access_token>` on every other call.
    - Access tokens last `expires_in` seconds (15 min). On `401`, call `POST /auth/refresh` once with the
...
      latest one. If refresh fails, send the user to login.
...

...

    ### Driver app: sending GPS
...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/otp/send:
    post:
      tags: [Auth]
      summary: Send a login OTP (Driver / Parent app)
      description: |
        Sends a 6-digit code by SMS, valid for `expires_in` seconds. The answer is the same whether or not
        the number is registered. Limits: one request per 30 s and 5 per hour per number.
      operationId: sendOtp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mobile, app]
              properties:
                mobile: { type: string, example: "9000000001" }
                app: { type: string, enum: [driver, parent] }
      responses:
        "200":
          description: Code sent (if the number is registered for that app).
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      sent: { type: boolean }
                      expires_in: { type: integer, description: seconds }
              example: { data: { sent: true, expires_in: 300 } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/v1/auth/otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP and log in (Driver / Parent app)
      description: A code works once. After 5 wrong attempts the code is cancelled; request a new one.
      operationId: verifyOtp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mobile, app, code]
              properties:
                mobile: { type: string, example: "9000000001" }
                app: { type: string, enum: [driver, parent] }
                code: { type: string, example: "123456" }
                device_name: { type: string, description: Shown in the user's session list., example: "Pixel 7" }
      responses:
        "200":
          description: Logged in.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Tokens" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Log in with email + password (admin web)
      description: For Super Admin, School Admin and Transport Manager. Drivers and parents use OTP.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                device_name: { type: string }
      responses:
        "200":
          description: Logged in, or a 2FA code is needed (`mfa_required`).
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/LoginResult" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/auth/login/2fa:
    post:
      tags: [Auth]
      summary: Complete login with an authenticator code
      operationId: login2fa
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mfa_token, code]
              properties:
                mfa_token: { type: string, description: From the login response. Valid 5 minutes. }
                code: { type: string, example: "492039" }
                device_name: { type: string }
      responses:
        "200":
          description: Logged in.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/LoginResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Get new tokens
      description: Returns a new access token **and a new refresh token**. The old refresh token stops working.
      operationId: refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: New tokens.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Tokens" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Log out this device
      operationId: logout
      responses:
        "204": { description: Session ended. }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current user and school
      description: Call after login and on app start.
      operationId: getMe
      responses:
        "200":
          description: The caller.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    required: [user, school, two_factor_setup_required]
                    properties:
                      user: { $ref: "#/components/schemas/User" }
                      school:
                        oneOf:
                          - $ref: "#/components/schemas/School"
                          - type: "null"
                        description: Null for Super Admin.
                      two_factor_setup_required: { type: boolean }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/v1/auth/sessions/{sessionID}:
    delete:
      tags: [Auth]
      summary: Log out one of my devices
      operationId: revokeSession
      parameters:
        - { name: sessionID, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Session ended. }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/auth/2fa/setup:
    post:
      tags: [Auth]
      summary: Start authenticator (2FA) setup (web roles)
      description: Returns a secret and QR code to scan. 2FA is not active until `/auth/2fa/enable` succeeds.
      operationId: setupTotp
      responses:
        "200":
          description: Scan the QR code with an authenticator app.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      secret: { type: string }
                      otpauth_url: { type: string }
                      qr_code_png: { type: string, description: "data: URL of a PNG" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/auth/2fa/enable:
    post:
      tags: [Auth]
      summary: Confirm a code and turn 2FA on
      operationId: enableTotp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
      responses:
        "204": { description: 2FA is on. Future logins need a code. }
        "400": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/schools:
    get:
      tags: [Schools]
      summary: List schools (Super Admin)
      operationId: listSchools
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
        - { name: q, in: query, description: Search name or code, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, inactive] } }
      responses:
        "200":
          description: Schools, by name.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/School" } }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Schools]
      summary: Create a school (Super Admin)
      operationId: createSchool
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SchoolInput" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/School" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/schools/{schoolID}:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
    get:
      tags: [Schools]
      summary: Get a school (Super Admin, or staff of that school)
      operationId: getSchool
      responses:
        "200":
          description: The school.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/School" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Schools]
      summary: Update a school (Super Admin)
      operationId: updateSchool
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SchoolInput" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/School" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/schools/{schoolID}/status:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
    patch:
      tags: [Schools]
      summary: Activate or deactivate a school (Super Admin)
      description: Users of an inactive school are logged out and cannot log in.
      operationId: setSchoolStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, inactive] }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/School" }
  /api/v1/schools/{schoolID}/users:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
    get:
      tags: [School users]
      summary: List School Admins and Transport Managers (Super Admin, School Admin)
      operationId: listSchoolUsers
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
        - { name: role, in: query, schema: { type: string, enum: [school_admin, transport_manager] } }
        - { name: status, in: query, schema: { type: string, enum: [active, suspended] } }
        - { name: q, in: query, description: "Search name, email or mobile", schema: { type: string } }
      responses:
        "200":
          description: Users, by name.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/User" } }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [School users]
      summary: Create a staff user
      description: Super Admin can create both roles; a School Admin can create Transport Managers only.
      operationId: createSchoolUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UserInput"
                - type: object
                  required: [role, password]
                  properties:
                    role: { type: string, enum: [school_admin, transport_manager] }
                    password: { type: string, minLength: 8 }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/schools/{schoolID}/users/{userID}:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { $ref: "#/components/parameters/UserID" }
    get:
      tags: [School users]
      summary: Get a staff user
      operationId: getSchoolUser
      responses:
        "200":
          description: The user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/User" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [School users]
      summary: Update name, email, mobile
      operationId: updateSchoolUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserInput" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/schools/{schoolID}/users/{userID}/status:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { $ref: "#/components/parameters/UserID" }
    patch:
      tags: [School users]
      summary: Suspend or reactivate
      description: Suspending logs the user out of every device.
      operationId: setSchoolUserStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, suspended] }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/User" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/schools/{schoolID}/users/{userID}/password:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { $ref: "#/components/parameters/UserID" }
    post:
      tags: [School users]
      summary: Set a new password
      description: Logs the user out of every device.
      operationId: resetSchoolUserPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 8 }
      responses:
        "204": { description: Password changed. }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/schools/{schoolID}/drivers:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
    get:
      tags: [Drivers]
      summary: List drivers
      operationId: listDrivers
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
        - { name: q, in: query, description: "Search name, mobile, licence", schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, inactive, suspended] } }
      responses:
        "200":
          description: Drivers, by name.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Driver" } }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Drivers]
      summary: Create a driver
      description: Also creates the driver's app login (mobile + OTP). The mobile must not belong to another driver.
      operationId: createDriver
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DriverInput" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Driver" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/schools/{schoolID}/drivers/{driverID}:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { name: driverID, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Drivers]
      summary: Get a driver
      operationId: getDriver
      responses:
        "200":
          description: The driver.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Driver" } } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Drivers]
      summary: Update a driver
      description: Changing the mobile changes the driver's login number.
      operationId: updateDriver
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DriverInput" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Driver" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/schools/{schoolID}/drivers/{driverID}/status:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { name: driverID, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [Drivers]
      summary: Change driver status
      description: Only an active driver can log in. Inactive or suspended drivers are logged out of every device.
      operationId: setDriverStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties: { status: { type: string, enum: [active, inactive, suspended] } }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Driver" } } }
  /api/v1/schools/{schoolID}/buses:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
    get:
      tags: [Buses]
      summary: List buses
      operationId: listBuses
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
        - { name: q, in: query, description: "Search vehicle number, make/model", schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, inactive, maintenance] } }
      responses:
        "200":
          description: Buses, by vehicle number.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Bus" } }
                  meta: { $ref: "#/components/schemas/PageMeta" }
    post:
      tags: [Buses]
      summary: Create a bus
      description: Vehicle numbers are unique per school, ignoring spaces and hyphens.
      operationId: createBus
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BusInput" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Bus" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/schools/{schoolID}/buses/{busID}:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { name: busID, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Buses]
      summary: Get a bus
      operationId: getBus
      responses:
        "200":
          description: The bus.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Bus" } } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Buses]
      summary: Update a bus
      operationId: updateBus
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BusInput" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Bus" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/schools/{schoolID}/buses/{busID}/status:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { name: busID, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [Buses]
      summary: Change bus status
      operationId: setBusStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties: { status: { type: string, enum: [active, inactive, maintenance] } }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Bus" } } }
  /api/v1/schools/{schoolID}/routes:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
    get:
      tags: [Routes & stops]
      summary: List routes
      description: Without stops; each route has `stop_count`.
      operationId: listRoutes
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
        - { name: q, in: query, description: Search name or code, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, inactive] } }
      responses:
        "200":
          description: Routes, by code.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Route" } }
                  meta: { $ref: "#/components/schemas/PageMeta" }
    post:
      tags: [Routes & stops]
      summary: Create a route
      operationId: createRoute
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RouteInput" }
      responses:
        "201":
          description: Created (with an empty `stops` list).
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Route" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/schools/{schoolID}/students/classes:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
    get:
      tags: [Students]
      summary: Distinct class values (for filters)
      operationId: listStudentClasses
      responses:
        "200":
          description: Classes in use.
          content:
            application/json:
              schema: { type: object, properties: { data: { type: array, items: { type: string } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
...
  responses:
...
    NotFound:
      description: Missing, or belongs to another school.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...